Replacement of Double to BigDecimal for more precision

+++ src/main/java/ch/bastiangardel/LittleCafet/dto/UserInfoDTO.java
@@ -2,6 +2,8 @@ package ch.bastiangardel.LittleCafet.dto;
 
 import ch.bastiangardel.LittleCafet.model.User;
 
+import java.math.BigDecimal;
+
 /**
  * Created by bastiangardel on 01.07.17.
  *
@@ -25,7 +27,7 @@ import ch.bastiangardel.LittleCafet.model.User;
 public class UserInfoDTO {
     private String username;
     private String name;
-    private double solde;
+    private BigDecimal solde;
 
     public UserInfoDTO modelToDTO(User user){
         username = user.getEmail();
@@ -42,11 +44,11 @@ public class UserInfoDTO {
         this.username = username;
     }
 
-    public double getSolde() {
+    public BigDecimal getSolde() {
         return solde;
     }
 
-    public void setSolde(double solde) {
+    public void setSolde(BigDecimal solde) {
         this.solde = solde;
     }
